# services/api/src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Hill90 API
  description: Hill90 platform API — agent management and lifecycle operations.
  version: 0.1.0
servers:
  - url: /
    description: Relative (behind Traefik reverse proxy)
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        agent_id:
          type: string
          description: Unique slug identifier (lowercase, alphanumeric, hyphens)
          pattern: "^[a-z0-9][a-z0-9-]{0,61}[a-z0-9]$"
        name:
          type: string
        description:
          type: string
        tools_config:
          type: object
        cpus:
          type: string
          default: "1.0"
        mem_limit:
          type: string
          default: "1g"
        pids_limit:
          type: integer
          default: 200
        soul_md:
          type: string
        rules_md:
          type: string
        status:
          type: string
          enum: [stopped, running, error, unknown]
          description: >-
            `unknown` means reconciliation could not verify this agent against a
            real container — the API declines to report the last value the
            database recorded as though it were current (#238). It is not a
            state an agent is ever put INTO; it is the API declining to answer.
        status_verified:
          type: boolean
          description: >-
            False when `status` is `unknown`. Carried separately so a consumer
            can keep the distinction without string-matching on the status.
        container_state:
          type: string
          nullable: true
          enum: [running, created, restarting, paused, exited, dead, absent, null]
          description: >-
            What the reconciler last SAW, as distinct from what the agent is
            recorded as (#239). `absent` means the container was not there at
            all — a 404 from the daemon, i.e. deleted rather than stopped —
            which is a different fact from `exited`. Null means the reconciler
            could not tell, and is NOT a synonym for `absent`.
        container_id:
          type: string
          nullable: true
        model_policy_id:
          type: string
          format: uuid
          nullable: true
          description: Internal policy reference used by backend model-router plumbing.
        model_policy_name:
          type: string
          nullable: true
          description: Human-readable name of the model policy (resolved via JOIN).
        models:
          type: array
          items:
            type: string
          description: User-facing model names assigned to this agent.
        skills:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              scope:
                type: string
                enum: [container_local, host_docker, vps_system]
              tools:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    name:
                      type: string
                    description:
                      type: string
                    install_method:
                      type: string
                      enum: [builtin, apt, binary]
          description: Skills assigned to this agent.
        container_profile_id:
          type: string
          format: uuid
          nullable: true
          description: FK to container_profiles table. Determines Docker image for the agent container.
        container_profile:
          type: object
          nullable: true
          description: Resolved container profile object (from LEFT JOIN).
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            docker_image:
              type: string
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
        principal_id:
          type: string
          format: uuid
          description: "Stable principal identifier (equals agent UUID). AI-115 workload principal model."
        principal_type:
          type: string
          enum: [agent]
          description: "Formal principal type. Always 'agent' for workload principals."
...
paths:
...
  /agents:
    get:
      summary: List agents
      description: Returns agents scoped to the authenticated user (all agents for admin).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
        "401":
          description: Not authenticated
        "403":
          description: Requires user role

    post:
      summary: Create agent
      description: Creates a new agent definition.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agent_id
                - name
              properties:
                agent_id:
                  type: string
                  pattern: "^[a-z0-9][a-z0-9-]{0,61}[a-z0-9]$"
                name:
                  type: string
                description:
                  type: string
                tools_config:
                  type: object
                cpus:
                  type: string
                mem_limit:
                  type: string
                pids_limit:
                  type: integer
                soul_md:
                  type: string
                rules_md:
                  type: string
                model_policy_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: Internal-only override. Prefer model_names for user-facing model assignment.
                model_names:
                  type: array
                  items:
                    type: string
                  description: User-facing model names to assign directly to this agent (one or many).
                skill_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Skill IDs to assign. Config is merged from all skills at save time (resolve-on-save).
                container_profile_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: Container profile to assign. Determines Docker image. Omit for default.
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Invalid input
        "401":
          description: Not authenticated
        "403":
          description: Requires user role
        "409":
          description: Duplicate agent_id
...